refactor: move load_config function to utilities module
- Relocated load_config() from main.py to utilities.py for better organization
- Updated import in main.py to use the function from its new location
- No functional changes were made to the function itself
